requirements.txt
alabaster==0.7.16
anyio==4.6.2.post1
appnope==0.1.3
argh==0.31.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
bump2version==0.5.11
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
convertdate==2.4.0
coverage==6.3.2
cvxpy==1.2.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker==7.1.0
docutils==0.17.1
ecos==2.0.14
entrypoints==0.4
ephem==4.1.6
exceptiongroup==1.2.2
fastjsonschema==2.20.0
flake8==6.0.0
flask==3.1.0
fonttools==4.55.0
fqdn==1.5.1
greykite==1.0.0
hijri-converter==2.3.1
holidays==0.13
holidays-ext==0.0.8
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.4.1
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-client==7.4.7
jupyter-console==6.4.4
jupyter-core==5.2.0
jupyter-events==0.10.0
jupyter-server==2.2.1
jupyter-server-terminals==0.5.3
jupyterlab-pygments==0.3.0
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
LunarCalendar-ext==0.0.1
MarkupSafe==3.0.2
matplotlib==3.5.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==2.0.5
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.4.0
nbformat==5.5.0
nest-asyncio==1.6.0
nh3==0.2.18
notebook==6.5.2
notebook-shim==0.2.4
numpy==1.23.2
osqp==0.6.2
overrides==7.3.1
packaging==24.2
pandas==1.5.0
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
patsy==0.5.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.11.2
platformdirs==4.3.6
plotly==5.4.0
pluggy==1.5.0
pmdarima==1.8.5
prometheus-client==0.21.0
prompt-toolkit==3.0.48
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
pygments==2.18.0
PyMeeus==0.5.12
pyparsing==3.2.0
pytest==7.1.2
pytest-runner==5.3.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.0.0
qdldl==0.1.7.post4
qtconsole==5.6.1
QtPy==2.4.2
readme-renderer==43.0
referencing==0.35.1
requests==2.28.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
scikit-learn==1.1.2
scipy==1.8.0
scs==3.2.7
seaborn==0.11.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
sphinx==5.3.0
sphinx-gallery==0.10.1
sphinx-rtd-theme==1.1.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
statsmodels==0.13.5
tenacity==9.0.0
terminado==0.18.1
testfixtures==6.14.2
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.14.3
twine==1.14.0
types-python-dateutil==2.9.0.20241003
typing-extensions==4.12.2
uri-template==1.3.0
urllib3==1.26.20
watchdog==0.9.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
widgetsnbextension==3.5.2
pyyaml
